Refactor XML macro handling code for use outside of tools.
